pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "snake-rl"
version = "0.1.0"
description = "Snake-from-Pixels Reinforcement Learning Project"
authors = ["Snake RL Team"]
license = "MIT"
readme = "README.md"
packages = [{include = "snake_rl"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.24.0"
torch = ">=2.1.0"
torchvision = ">=0.16.0"
gymnasium = "^0.29.0"
pygame = "^2.5.0"
tensorboard = "^2.15.0"
wandb = "^0.16.0"
hydra-core = "^1.3.0"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
tqdm = "^4.66.0"
omegaconf = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"
flake8 = "^6.0.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["snake_rl"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pygame.*",
    "wandb.*",
    "tensorboard.*",
    "cv2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["snake_rl/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=snake_rl",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["snake_rl"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]